% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_model_type.R
\name{detect_model_type}
\alias{detect_model_type}
\title{Detect Model Type}
\usage{
detect_model_type(data, response)
}
\arguments{
\item{data}{A data frame.}

\item{response}{The name of the response variable.}
}
\value{
A character string: "linear" or "logistic".
}
\description{
Automatically detects whether the response variable should be modeled as
a linear or logistic regression based on its type.
}
