% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_plausible_models().R
\name{plausible_models}
\alias{plausible_models}
\title{Plausible Model Selection with AIC and Stability Filtering}
\usage{
plausible_models(full_data_models, stability_df, delta = 2, tau = 0.6)
}
\arguments{
\item{full_data_models}{List of models from multi_path_forward().}

\item{stability_df}{Tibble of stability scores from stability_resampling().}

\item{delta}{Numeric; maximum AIC difference from best model.(default 2)}

\item{tau}{Numeric; minimum average stability threshold.(default 0.6)}
}
\value{
A List ("plausible_models") of plausible and stable models that contain:
\describe{
 \item{fit}{A fitted model}
 \item{aic}{AIC value for the model}
 \item{vars}{Predictor variable names within models}
 \item{avg_stability}{Average stability score of the predicitor models}
}
}
\description{
Selects plausible and stable models using AIC and average stability thresholds
}
