% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stability.R
\name{plot_stability}
\alias{plot_stability}
\title{Plot Variable Stability}
\usage{
plot_stability(stability_df)
}
\arguments{
\item{stability_df}{Tibble from stability_resampling().}
}
\description{
Visualizes stability scores as a bar plot.
}
