% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability().R
\name{stability}
\alias{stability}
\title{Stability Resampling}
\usage{
stability(
  X,
  response,
  B = 100,
  K = 5,
  epsilon = 1e-06,
  delta = 2,
  L = 25,
  model_type = NULL
)
}
\arguments{
\item{X}{Data frame of predictors and response.}

\item{response}{is the name of the response variable}

\item{B}{Number of bootstrap samples.}

\item{K}{Maximum number of forward steps (default 5).}

\item{epsilon}{Minimum AIC improvement to continue from a parent (default 1e-6)}

\item{delta}{AIC tie threshold; keep children within `best + delta` (default 2)}

\item{L}{Maximum number of models to keep per step (default 25)}

\item{model_type}{Model type is specified ("linear" or "logistic"). Automatically detected if not provided ("NULL").}
}
\value{
A dataframe 'path_stability' with two columns
\describe{
 \item{variable}{predictor names}
 \item{pi}{The stabilty score for each variable}
}
}
\description{
This function estimates the stability of variable selection using repeated bootstrap method.
}
\examples{
# Stability function for a linear response
Linear_scores <- stability(X = mtcars, response = "mpg", K = 3, delta = 2)
# Stability function for a logistic response
mtcars$am <- factor(mtcars$am) #convert to factor
Logistic_scores <- stability(X = mtcars, response = "am", K = 3, delta = 2)
}
